Move refresh from timeseries to actions
