+++ src/services/useTimeSeries/index.ts
@@ -14,24 +14,20 @@ import { createTransformRequestFn } from '@/lib/requests/transformRequest'
 import { difference } from 'lodash-es'
 import { SeriesData } from '@/lib/timeseries/types/SeriesData'
 import { convertFewsPiDateTimeToJsDate } from '@/lib/date'
-import { type Pausable } from '@vueuse/core'
-import { useFocusAwareInterval } from '@/services/useFocusAwareInterval'
 
 export interface UseTimeSeriesReturn {
   series: ShallowRef<Record<string, Series>>
   isLoading: Ref<boolean>
   loadingSeriesIds: Ref<string[]>
-  interval: Pausable
   refresh: () => void
 }
-
-const TIMESERIES_POLLING_INTERVAL = 1000 * 30
 
 export interface UseTimeSeriesOptions {
   startTime?: Date | null
   endTime?: Date | null
   thinning?: boolean
   showVerticalProfile?: boolean
+  enabled?: boolean
 }
 
 function timeZoneOffsetString(offset: number): string {
@@ -68,6 +64,8 @@ export function useTimeSeries(
     const _requests = toValue(requests)
     const _options = toValue(options)
     const _selectedTime = toValue(selectedTime)
+
+    if (!_options?.enabled) return
 
     const currentSeriesIds = Object.keys(series.value)
     const updatedSeriesIds: string[] = []
@@ -194,12 +192,6 @@ export function useTimeSeries(
     }
   }
 
-  const interval = useFocusAwareInterval(
-    loadTimeSeries,
-    TIMESERIES_POLLING_INTERVAL,
-    { immediateCallback: true },
-  )
-
   onUnmounted(() => {
     controller.abort('useTimeSeries unmounted.')
   })
@@ -208,7 +200,6 @@ export function useTimeSeries(
     series,
     isLoading,
     loadingSeriesIds,
-    interval,
     refresh: loadTimeSeries,
   }
 }
